requirements.txt
matplotlib
pandas
numpy
seaborn
scipy
tabulate
beautifulsoup4
datetime==5.5
fpdf==1.7.2
mplsoccer==1.4.0
rembg==2.0.62
pillow==11.1.0
requests== 2.32.3
imageio==2.37.0
streamlit==1.41.1
onnxruntime
ollama
